Add support for pointer navigation and fire control
It's difficult to move precisely using the arrow keys. We add support
for point-and-click navigation so that the user can click a target and
move to it. There are still some things to work out surrounding wall
collisions, but it works for the most part.

We also add support for throwing the knife in the direction of the mouse
pointer. A side-effect of the implementation is that if the pointer is
used to move the player, the knives are thrown in the same direction.

The old style of navigation/aiming has been maintained and will override
the new style.