Overlay fireball with particles
We add associate a particle emitter with each fireball when it is
thrown. The emitter follows the fireball and the fireball is the index
to retrieve the emitter when it's time to handle collisions.

At the moment, the fireball is only for display and goes through walls.
Also of note, wizards without line-of-sight can still cast a fireball at
the player.