Don't let wizards see through walls
Previously, a wizard only had to be vertically or horitontally aligned
in order to fire upon a player. This doesn't make sense if there's a
wall blocking them. We now pass the walls layer to the wizard and
consider the player out of site if a collidable wall tile is
obstructing.