Move most methods out of Player class
Right now we're using inheritance to extend
`Phaser.Physics.Arcade.Sprite` to tha `Player` is a `Sprite`. We want
to change that relationship so that `Player` has a `Sprite`. This makes
it easier to think about the sprite in isolation and makes it easier to
decouple our framework from our business logic.