Make new dungeon
We want to make a Legend of Zelda-like dungeon with discrete rooms which
we transition between. The first step is create a new tilemap with the
rooms we need.

Each room has an invisible object which holds metadata (such as "name")
for that room. The first room also has a door object with a destination
which determines where the door leads to.

The hope is that by associating a door's destination with a room's
name, we can pan the camera to the new room's center.