Allow player to open door
We add a collision handler which delegates to a function on the player
setting the door as an active door for the player. We only want the
handler to run when the door is closed so we pass in a function to let
the handler know this.

We also open a door when the player uses the action button and there's
an active door.