Detect when the player enters and exits a door
In order to leave a room and enter another we need keep track of which
door was entered, which direction the player was heading when they
entered the door, the direction the player is currently facing, and
when the player exits the door.

We track this data, but don't perform an action when the player exits
a door.