Fix room transition
We add metadata to the tilemap which we interpret as a graph of door and
room relationships during initialization. We use this graph to determine
which room the camera should pan to when the player passes through a
door.

We also implement the door as a node in the aforementioned graph.