Make movement more abstract
We remove the move event which caused enemies to change direction
periodically and instead deal with enemy actions in the `update` method.

Now, we should be able to override the default `attack` method to change
an enemy's attack strategy.

We also prevent enemies from attacking if the player is not visible.