Parameterize knockBack
We don't want the boss to run away when attacked by the player, so we
allow it to set a negative `knockBack`. We could have dealt with this by
overriding `handleDamage`, but that's a lot more work than we need to do
right now. If we need to do that in the future we can revert this
change.