Flip Door image during initialization.
Previously, we flipped the door image when it was opened (since that's
the only time we noticed an issue). While testing, however, we
discovered that this results in an upside-down door in Level 2.

Now, we flip the door during initialization. We only do if for
south-facing doors since we only have north and south facing doors at
the moment.